+++ src/modules/admin/widgets/forms/fields/AssetPreview.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace davidhirtz\yii2\media\modules\admin\widgets\forms\fields;
+
+use davidhirtz\yii2\media\models\AssetInterface;
+use yii\base\Widget;
+use yii\helpers\Html;
+
+class AssetPreview extends Widget
+{
+    public ?AssetInterface $asset = null;
+
+    public function run()
+    {
+        $file = $this->asset->file;
+
+        if ($file->hasPreview()) {
+            $image = Html::img($file->getUrl(), [
+                'id' => 'image',
+                'class' => 'img-transparent',
+            ]);
+
+            if ($width = $file->width) {
+                $image = Html::tag('div', $image, [
+                    'style' => "max-width:{$width}px",
+                ]);
+            }
+
+            return $image;
+        }
+
+        return '';
+    }
+}